refactor(infra): move pulumi installs to root + rename infra/ => infra-cdk + rename infra-pulumi => infra
