+++ infra-cdk/lambdas/slack/package.json
