+++ infra-cdk/lambdas/slack/tests/sns-ok.json
