--- infra-pulumi/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "shepherd-infra",
-  "version": "1.0.0",
-  "type": "module",
-  "devDependencies": {
-    "@types/node": "^22.13.10",
-    "tsx": "^4.19.3",
-    "typescript": "^5.8.2"
-  },
-  "dependencies": {
-    "@pulumi/docker": "^4.11.2",
-    "@pulumi/minio": "^0.16.8",
-    "@pulumi/pulumi": "^3.0.0"
-  }
-}
