+++ infra/elasticmq/generate-config.ts
