+++ infra/minio2mq/Dockerfile
